cb_dist module to work with environment variables.
When the distribution is started, the listen callback of the cb_dist
module will get called. In this callback, we start the listeners based
on "local_preferred_dist" and "global_preferred_dist" environment
variables.

"local_preferred_dist" is the dist protocol used to talk to the local
erlang VMs (babysitter and couchdb).

"global_preferred_dist" is the dist protocol used to talk to the
remote ns_server nodes.