Allow net_kernel to be restarted when there is...
...a change in the networking configuration.

While handling the 'setupNetConfig' API, we will asynchronously notify
the babysitter VM to restart its children. The babysitter VM will
first update the environment variable with the new distribution type,
restart its net_kernel and then restart all its children.

This will cause the respective node to appear as offline to the other
nodes in the cluster resulting in an auto-failover. This is undesirable
as administrator will be aware that the network config update will
cause the nodes to go offline briefly. So we will fail the
setupNetConfig API if auto-failover is enabled in the cluster. The
administrator will have to disable auto-failover, update net config on
all the nodes in the cluster and when the cluster starts operating in
the new mode, the administrator will have to re-enable auto-failover.