Add ability to turn on/off cluster...
...encryption via REST API.

Adding a new parameter called 'clusterEncryption' to the existing
'/node/controller/setupNetConfig' REST API to enable node to node
encryption. The values this parameter accepts are: 'on' & 'off'.

When set to 'on', ns_server stores either 'inet_tls' or 'inet6_tls'
as the future global proto_dist type in the dist_cfg file. The local
proto_dist type will be set to either 'inet_tcp' or 'inet6_tcp'. The
reason is that the erlang distribution between local nodes need not
be encrypted.

As before this API needs to be called on every node in the cluster
and the REST API will instruct babysitter to stop ns_server, restart
its own net_kernel and then start ns_server back again for the change
to take effect.

For the TLS distribution to work correctly, Erlang stipulates the
presence of a custom path verification function. This has been defined
in the dist_manager module.

When operating in TLS distribution mode, the 'ssl_dist_optfile'
parameter must also be passed down to the couchdb VM, so explictly
pass it on to the couchdb VM.

Note that the 'ssl_dist_optfile' parameter will always be passed while
starting or stopping the Couchbase server in the init script and
cluster_run. This parameter will be ignored when operating in non-TLS
mode but will be used internally by erlang's 'inet(6)_tls_dist' module
to fetch the SSL options.